requirements.txt
fastapi>=0.110
uvicorn>=0.30
jinja2>=3.1
python-multipart>=0.0.9
ultralytics
torch
torchvision
opencv-python
numpy
matplotlib
tqdm
umap-learn
